requirements.txt
fastapi==0.115.12
Markdown==3.7
openai==1.82.0
pandas==2.2.2
pydantic==2.10.6
python-dotenv==1.1.1
requests==2.32.3
uvicorn==0.34.2
pydantic==2.10.6
jinja2==3.1.6
tabulate==0.9.0
python-multipart==0.0.20